Another guint.

+++ tests/libpay-tests.cpp
@@ -259,7 +259,7 @@ TEST_F(LibPayTests, ItemLifecycle)
     dbus_test_dbus_mock_object_emit_signal(mock, pkgobj,
                                            "ItemStatusChanged",
                                            G_VARIANT_TYPE("(sst)"),
-                                           g_variant_new("(sst)", "item", "purchasing", 0),
+                                           g_variant_new("(sst)", "item", "purchasing", (guint64)0),
                                            &error);
     EXPECT_EQ(nullptr, error);
